long->int32

+++ src/eg3dmath.c
@@ -24,7 +24,7 @@ static int16 sineInterp(uint16 angle) {
     int16 frac = angle & 0xFF;
     int16 v0 = g_angleLut[idx];
     int16 v1 = g_angleLut[idx + 1];
-    long step = (long)(v1 - v0) * frac;
+    int32 step = (int32)(v1 - v0) * frac;
     return v0 + (int16)((step + 0x80) >> 8);
 }
 
@@ -41,17 +41,17 @@ int16 cosine(int16 angle) {
  * Replicates the exact shl/rcl/adc sequence of the ASM fixedMulQ14:
  *   P = a*b; result = (P>>15) + (bit14 of P). */
 int16 fixedMulQ14(int16 a, int16 b) {
-    long p = (long)a * (long)b;
+    int32 p = (int32)a * (int32)b;
     return (int16)((p >> 15) + ((p >> 14) & 1L));
 }
 
 /* In-place 32-bit shifts (pascal: args pushed left-to-right, callee cleans up).
- * The ASM dispatches to the MSC long-shift helpers; >>= on a signed long is the
+ * The ASM dispatches to the MSC long-shift helpers; >>= on a signed int32 is the
  * arithmetic shift those helpers perform. */
-void pascal shiftLongLeftInPlace(int16 count, long *ptr) {
+void pascal shiftLongLeftInPlace(int16 count, int32 *ptr) {
     *ptr <<= count;
 }
 
-void pascal shiftLongRightInPlace(int16 count, long *ptr) {
+void pascal shiftLongRightInPlace(int16 count, int32 *ptr) {
     *ptr >>= count;
 }
